+++ R/init.R
@@ -1,8 +1,7 @@
 #' Initialise site infrastructure
 #'
-#' This creates the output directory (`docs/`), `favicon.ico` (from the package
-#' logo), a machine readable description of the site, and copies CSS/JS
-#' assets and extra files.
+#' This creates the output directory (`docs/`), a machine readable description
+#' of the site, and copies CSS/JS assets and extra files.
 #'
 #' @section Build-ignored files:
 #' We recommend using [usethis::use_pkgdown()] to build-ignore `docs/` and
@@ -18,9 +17,9 @@
 #' `<HEAD>` after the default pkgdown CSS and JS.
 #'
 #' @section Favicon:
-#' If you include you package logo in the standard location of
-#' `man/figures/logo.png`, a favicon will be automatically created for
-#' you.
+#' You should manually run [build_favicon()] once to generate the favicon set
+#' from your logo. The result is stored in `pkgdown/favicon` and will
+#' automatically be copied to the relevant location when you run [init_site()].
 #'
 #' @inheritParams build_articles
 #' @export