+++ NEWS.md
@@ -1,5 +1,11 @@
 # pkgdown 1.1.0.9000
 
+* New function `build_favicon()` creates high resolution favicons, which are 
+  better suited for modern web usage (e.g. retina display screens, desktop
+  shortcuts, etc.).  For this, it uses the <http://realfavicongenerator.net>
+  API. As a side effect, this change also removes the dependency to the
+  magick package (@bisaloo, #883).
+
 * Navbar version now gets class "version" so you can more easily control the
   display if you want (#680).